requirements.txt
streamlit==1.32.0

# LangChain ecosystem - pinned to stable compatible versions
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-openai==0.2.11
langchain-anthropic==0.3.1
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.3

# Vector store & embeddings
faiss-cpu==1.9.0
sentence-transformers==3.3.1

# PDF processing
pypdf==5.1.0
PyPDF2==3.0.1

# Environment & utilities
python-dotenv==1.0.1
tiktoken==0.8.0

# Data & visualization
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.4
seaborn==0.13.2
fpdf2==2.8.1

# Optional requirements for QLoRA and LoRA (Better PDF Reader Local Extraction)
# If using a GPU machine, uncomment these lines:
# torch>=2.0.0
# transformers>=4.38.2
# accelerate>=0.28.0
# bitsandbytes>=0.43.0
# peft>=0.9.0
